Some corrections to de.json
Added one missing s.

Found another personal pronoun and converted it to lowercase.

Corrected labeling of the Load Plan Icon to "Laden" (Load) instead of the confusing naming "Gespeichert" (Saved).
The opposite Icon is correctly named "Speichern" (Save).

The Reset Plan Icon is named "Clear" in english. Clear means "Löschen" but not "Zurücksetzen" (which means Reset).
The Restart Plan Icon is named "Restart" in english. Restart is a noun, so "Neustart" is the correct translation instead of "Neu starten".