Update hu.json (#904)
I systematically searched the untranslated texts on the main window and on the settings windows. 
I translated all the texts that appears on the main window of the app, the titles, descriptions, hidden info texts related to the main and the different setting windows.
Hopefully this will help the newer users using Hungarian as the application's preferred Language.

After this package I would like to make a following package, wich would comprise the texts used by voice navigation.